리소스 파일 정리 / HUD 제거 / Test 컨트롤러 추가,삭제
