--- Content/ProjectEscape/UI/4343.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ce28262546c60e524f01d931aab7384d025028d869062558e277aa900e716602
-size 27175
