--- Content/ProjectEscape/UI/HUD/BP_PEHUD.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:01f4f0518672625eeb42f38d5ef3001a6afd8e91c2d83372cf627685c85573dd
-size 23817
