+++ Content/ProjectEscape/UI/HUD/WBP_Damage.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d2a89fec7eda09cfeecbc2490a2031274badd01f4ab69a8bb6f85c461fd64f6e
+size 55276
