+++ Content/ProjectEscape/UI/HUD/WBP_HUD.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:651527d580e3980256a44695ea0c05d2fc2c3599eea9e67a1c38ea8a67687ac2
-size 54842
+oid sha256:e07da9490493ec5e9b448d52dad6d73af242fa119b6ff302efe2fcb52bbb4c94
+size 54686
