--- Content/ProjectEscape/UI/HealthBar-ex.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4e19924575553422ce3bd6733ece105c7adc4486a40e6b61aeabcdc31d74c22a
-size 14028
