--- Content/ProjectEscape/UI/HealthBar-ex3.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:32e3d91c68a62dc17c466cf6888ebf9860ae47f30b178f24f5375b9730f1cfa2
-size 20250
