+++ Content/ProjectEscape/UI/Menu/WBP_Applications_BigButton.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9fa8460f3250d6cbda43331884ce6d38ebb178d9239a4f7bd21589e118cebc7f
-size 91433
+oid sha256:f9d2cdb56d2233581f125845f5d631536236b3c8947b31607ead1e8db6428aa4
+size 91393
