+++ Content/ProjectEscape/UI/Menu/WBP_MainMenu.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:eaa273ccfb6e5bebb939294bfed48c3ff25d16558fea89e09407e73c876faae7
-size 60570
+oid sha256:7008637aabfdd1017ea15f26d25ef3baf12a4462381d333569b562f6c6f5e993
+size 60519
