+++ Content/ProjectEscape/UI/Menu/WBP_PauseMenu.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:10c26f2e2e39eebf7b12f26c2ede57233f40b24759e3987695397fb3d6eff6c9
-size 60511
+oid sha256:860f6ea91ee920551a2e3a8065bab4974f60078b87a5584b14d793ddff8012f3
+size 60455
