--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Distortions/MF_UI_Slant.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:692af02c009fc56bfa0b36daf23a65ca9c2911d6fbda03e528d8066b4e455afa
-size 17797
