--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/SDFs/MF_UI_SDF_Box.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3a99e0c8c4ee42fdae998854bae77d13c391e72f2e02569af85d440b08ab66f0
-size 56595
