--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/SDFs/MF_UI_SDF_Hexagon.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:dc5a40ae488cfd0fb6951c8014cd6de9d72f37224f61afc050b74a3d30a271a1
-size 55745
