--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/SDFs/MF_UI_Wave.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4cbbf201ea7e667c512f242ad42a0fdcb443b3e75432d7093de77f183cb22d20
-size 28225
