--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Time/MF_UI_LinearTime.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:21207802046b1bec5528db61be0e690e6440ff9b303613704769f7d67b4ee5a0
-size 16928
