--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Transforms/MF_UI_Translate.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:74f48c4416b09d0f4d27ba1fdec56db9bcd9a7a0d66ad45777f0c01e635224f9
-size 14638
