--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_AspectRatio.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:37cac9641f3f7661db41eec0d7d086041296bbc51ce67bee45c028264c7e6727
-size 25785
