--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_EaseCurves.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7eed41732c4e3f7e083f39b6083070d529776a296ae5245620d9d0aee1ccd2a1
-size 53970
