--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_PingPong.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9acf677522cde6bd972964671dda853efc0b575f68f86c19b3974d9f83b6b6b1
-size 24031
