--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_RemappableGradient.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0caae100a6b24ae77c5074a449aab68d5ec4bd31185563207054738bd2354067
-size 47489
