--- Content/ProjectEscape/UI/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_WeightedCorners.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d6ee97d54d2ef369870bbe3f14082f721defe68635df22091c38b870cae58d3c
-size 29962
