--- Content/ProjectEscape/UI/UIMaterialLab/Widgets/ComponentMaterials/M_UI_Button3.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:99167e78ec1bdb58c8524b077a46f651f0a19d380a92b707dd95534ce4660000
-size 91631
