--- Content/ProjectEscape/UI/UIMaterialLab/Widgets/ComponentMaterials/MaterialInstances/MI_UI_Button3.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:dbe7e4d21497f8f622eefec09dfe79742cc92f46c1ed50bb0e9b3c7dcc7a8373
-size 18576
