--- Content/ProjectEscape/UI/UIMaterialLab/Widgets/Components/WBP_Applications_Button_3.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:bd0a6ed37c5768b52f0dbb098955725cf18ed6baaf2741317a8326eac90bb093
-size 88522
