--- Content/ProjectEscape/UI/WBP_Damage.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:dbb3beadbb93c12af335883aa4c00261e793b7be9b4a5a404ccece05f77b1b2f
-size 55256
