--- Content/ProjectEscape/UI/crosshair.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:37a1f6b16490de8651c22f4886f9bbf167d309bf0508c72f52987abcc24ee89a
-size 12847
