--- Content/ProjectEscape/UI/gunet.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:60f47937d25f2e72409f1512be115a0925c0c8b1de1fed79bfee179162770beb
-size 17470
