--- Content/ProjectEscape/UI/weapon-soket.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5dc563c458623674dbd945ce16bc88177d45aff0bb59ef1fca65e2e2ba625c7a
-size 18976
