--- Content/ProjectEscape/UI/weapon.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:198183518fcd20746834784aaef0a6316c16a1f69c8d090c0c82c437bd489b78
-size 44167
