+++ Content/Resources/UIResources/4343.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:bc19ab54d7af8dc9b4085010dcd6b96b4a101161dc3530580b0cdbef40fed8e5
+size 27857
