+++ Content/Resources/UIResources/HealthBar-ex.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:da268d64b531d75b19b332db1e922367392f3ab7277327f979ce67b0f5beaf7a
+size 14710
