+++ Content/Resources/UIResources/HealthBar-ex3.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:aa878b3a4f4b22f8c2633ff928dea99a0bf0aaf0f5736e2f7a84f75ac7a791b4
+size 20932
