+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/Distortions/MF_UI_Slant.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:76824a2e87f1c84444274e79c10821a81db8fda85d08630937fdf97e79e656b3
+size 17812
