+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/SDFs/MF_UI_SDF_Box.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e9da21bc2ed4a8bd9232ef301c170a8497ff40399ceac9bf088d0453f5c03fae
+size 56620
