+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/SDFs/MF_UI_SDF_Hexagon.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:05f7512ab7031c5d8ba261614987ee030ee0e41abeebb3f3d59e00395e90ba40
+size 55760
