+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/SDFs/MF_UI_Wave.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ba4b6afebe8d3b979b72ed50673154de4f230102e71094a950341207d8801bc2
+size 28250
