+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_AspectRatio.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4e6e73076e2088212346e698ce463d5713cb9ee5b0531a5e90549e42bc51131d
+size 25800
