+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_EaseCurves.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c134457dbb6d56302bb14f95c1ff5dbab454efd33daad7bb804092612b382507
+size 53990
