+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_PingPong.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fbd27fc47323d75b6d4a0847659cb8f4fd011b88350b5ac8a78be96f5c5a0c12
+size 24046
