+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_RemappableGradient.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b4381b2f193292b5e03f51994512864bff9dfbc7a663c02423d5f0c6f4e14457
+size 47504
