+++ Content/Resources/UIResources/UIMaterialLab/MaterialFunctions/Utilities/MF_UI_WeightedCorners.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2bddf470bd2f8191b472b03ca273436f7b56213f41b5e7820c1513f503a143a9
+size 29977
