+++ Content/Resources/UIResources/UIMaterialLab/Widgets/ComponentMaterials/M_UI_Button3.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0cd7dcccfb5d121e682393c685ae6d51aaa838b227f924ca944f4599954f4521
+size 91696
