+++ Content/Resources/UIResources/UIMaterialLab/Widgets/ComponentMaterials/MaterialInstances/MI_UI_Button3.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1cead61abff2fee0287fd735378f56fc232f471b98fefe92246be22771de854b
+size 18596
