+++ Content/Resources/UIResources/UIMaterialLab/Widgets/Components/WBP_Applications_Button_3.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:01c4ecf1c3e5db956472893f82a211ad184e3d2b5ad0e495d833c038062d564d
+size 88552
