+++ Content/Resources/UIResources/crosshair.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:176a5c1b923c994a25d22b085d3357bfce859b6a07e851398553f4d15ea9d974
+size 13529
