+++ Content/Resources/UIResources/weapon.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c7236372233073088e14002ba69af8008f2998d2a899c7eb9fd90d555b864787
+size 44849
