--- Content/Test/BasicUITest/BP_PEUIPlayerController.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e4418995babef974af57328b00a99ebdc8a433a2214dee6e88b3184d87efd8b7
-size 23153
