+++ Content/Test/BasicUITest/UITestMap_Simple.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:ad8ccb2862b7b39cb4ec67bd227afefe4e8f8152078af8301a3c5a323a76523a
+oid sha256:7c8faaf69c68726ba5c1cb973f5904a5061b3cd20b58af85f5b2394b5013590c
 size 9262
