+++ Content/Test/BasicUITest/UITest_Basic_GameModeBase.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1b5aaabb3688ab56b29fb0d682bfedc2072b688ae2d2d2c5c1b7edfb59473731
-size 20817
+oid sha256:f0b962d6bd5254253821a5792c78f5884cca4431079a894d35861e0db51a53a1
+size 21088
